# auth/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Auth system made for kyoo.",
        "title": "Keibi - Kyoo's auth",
        "contact": {
            "name": "Repository",
            "url": "https://github.com/zoriya/kyoo"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "1.0"
    },
    "host": "kyoo.zoriya.dev",
    "basePath": "/auth",
    "paths": {
...
        "/jwt": {
            "get": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Convert a session token or an API key to a short lived JWT.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jwt"
                ],
                "summary": "Get JWT",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Jwt"
                        },
                        "headers": {
                            "Authorization": {
                                "type": "string",
                                "description": "Jwt (same value as the returned token)"
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid session token (or expired)",
                        "schema": {
                            "$ref": "#/definitions/main.KError"
                        }
                    }
                }
            }
        },
        "/keys": {
            "get": {
                "security": [
                    {
                        "Jwt": [
                            "apikeys.read"
                        ]
                    }
                ],
                "description": "List all api keys",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apikeys"
                ],
                "summary": "List API keys",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Page-main_ApiKey"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Jwt": [
                            "apikeys.write"
                        ]
                    }
                ],
                "description": "Create a new API key",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apikeys"
                ],
                "summary": "Create API key",
                "parameters": [
                    {
                        "description": "Api key info",
                        "name": "key",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/main.ApiKeyDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.ApiKeyWToken"
                        }
                    },
                    "409": {
                        "description": "Duplicated api key",
                        "schema": {
                            "$ref": "#/definitions/main.KError"
                        }
                    },
                    "422": {
                        "description": "Invalid create body",
                        "schema": {
                            "$ref": "#/definitions/main.KError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Jwt": [
                            "apikeys.write"
                        ]
                    }
                ],
                "description": "Delete an existing API key",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apikeys"
                ],
                "summary": "Delete API key",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.ApiKey"
                        }
                    },
                    "404": {
                        "description": "Invalid id",
                        "schema": {
                            "$ref": "#/definitions/main.KError"
                        }
                    },
                    "422": {
                        "description": "Invalid id format",
                        "schema": {
                            "$ref": "#/definitions/main.KError"
                        }
                    }
                }
            }
        },
        "/oidc/callback/{provider}": {
            "get": {
                "description": "Exchange an opaque OIDC token for a local session.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oidc"
                ],
                "summary": "OIDC callback",
                "parameters": [
                    {
                        "type": "string",
                        "example": "google",
                        "description": "OIDC provider id",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Opaque token returned by /oidc/logged/:provider",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Optional tenant passthrough for federated setups",
                        "name": "tenant",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Bearer token to link provider to current account",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.SessionWToken"
                        }
                    },
                    "404": {
                        "description": "Unknown OIDC provider",
                        "schema": {
                            "$ref": "#/definitions/main.KError"
                        }
                    },
                    "410": {
                        "description": "Login token expired or already used",
                        "schema": {
                            "$ref": "#/definitions/main.KError"
                        }
                    }
                }
            }
        },
        "/oidc/logged/{provider}": {
            "get": {
                "description": "Callback endpoint called by OIDC providers after login.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oidc"
                ],
                "summary": "OIDC logged callback",
                "parameters": [
                    {
                        "type": "string",
                        "example": "google",
                        "description": "OIDC provider id",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "State value returned by the provider",
                        "name": "state",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization code",
                        "name": "code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Provider callback error",
                        "name": "error",
                        "in": "query"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    },
                    "400": {
                        "description": "Invalid state",
                        "schema": {
                            "$ref": "#/definitions/main.KError"
                        }
                    },
                    "404": {
                        "description": "Unknown OIDC provider",
                        "schema": {
                            "$ref": "#/definitions/main.KError"
                        }
                    }
                }
            }
        },
        "/oidc/login/{provider}": {
            "get": {
                "description": "Start an OIDC login with a provider.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oidc"
                ],
                "summary": "OIDC login",
                "parameters": [
                    {
                        "type": "string",
                        "example": "google",
                        "description": "OIDC provider id",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "URL to redirect the browser to after provider callback",
                        "name": "redirectUrl",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Optional tenant passthrough for federated setups",
                        "name": "tenant",
                        "in": "query"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    },
                    "400": {
                        "description": "Missing redirectUrl",
                        "schema": {
                            "$ref": "#/definitions/main.KError"
                        }
                    },
                    "404": {
                        "description": "Unknown OIDC provider",
                        "schema": {
                            "$ref": "#/definitions/main.KError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Jwt": []
                    }
                ],
                "description": "Remove an OIDC provider from the current account.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oidc"
                ],
                "summary": "OIDC unlink provider",
                "parameters": [
                    {
                        "type": "string",
                        "example": "google",
                        "description": "OIDC provider id",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Unknown OIDC provider",
                        "schema": {
                            "$ref": "#/definitions/main.KError"
                        }
                    }
                }
            }
        },
        "/sessions": {
            "get": {
                "security": [
                    {
                        "Jwt": []
                    }
                ],
                "description": "List all active sessions for the currently connected user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "List my sessions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.SessionWCurrent"
                            }
                        }
                    },
                    "401": {
                        "description": "Missing jwt token",
                        "schema": {
                            "$ref": "#/definitions/main.KError"
                        }
                    },
                    "403": {
                        "description": "Invalid jwt token (or expired)",
                        "schema": {
                            "$ref": "#/definitions/main.KError"
                        }
                    }
                }
            },
            "post": {
                "description": "Login to your account and open a session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "type": "string",
                        "example": "android tv",
                        "description": "The device the created session will be used on",
                        "name": "device",
                        "in": "query"
                    },
                    {
                        "description": "Account informations",
                        "name": "login",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/main.LoginDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.SessionWToken"
                        }
                    },
                    "403": {
                        "description": "Invalid password",
                        "schema": {
                            "$ref": "#/definitions/main.KError"
                        }
                    },
                    "404": {
                        "description": "Account does not exists",
                        "schema": {
                            "$ref": "#/definitions/main.KError"
                        }
                    },
                    "422": {
                        "description": "User does not have a password (registered via oidc, please login via oidc)",
                        "schema": {
                            "$ref": "#/definitions/main.KError"
                        }
                    }
                }
            }
        },
...
        "/sessions/{id}": {
            "delete": {
                "security": [
                    {
                        "Jwt": []
                    }
                ],
                "description": "Delete a session and logout",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Delete other session",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "example": "e05089d6-9179-4b5b-a63e-94dd5fc2a397",
                        "description": "The id of the session to delete",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Session"
                        }
                    },
                    "404": {
                        "description": "Session not found with specified id (if not using the /current route)",
                        "schema": {
                            "$ref": "#/definitions/main.KError"
                        }
                    },
                    "422": {
                        "description": "Invalid session id",
                        "schema": {
                            "$ref": "#/definitions/main.KError"
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "security": [
                    {
                        "Jwt": [
                            "users.read"
                        ]
                    }
                ],
                "description": "List all users existing in this instance.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "List all users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "used for pagination.",
                        "name": "after",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Page-models_User"
                        }
                    },
                    "422": {
                        "description": "Invalid after id",
                        "schema": {
                            "$ref": "#/definitions/main.KError"
                        }
                    }
                }
            },
            "post": {
                "description": "Register as a new user and open a session for it",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Register",
                "parameters": [
                    {
                        "type": "string",
                        "example": "android",
                        "description": "The device the created session will be used on",
                        "name": "device",
                        "in": "query"
                    },
                    {
                        "description": "Registration informations",
                        "name": "user",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.RegisterDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.SessionWToken"
                        }
                    },
                    "403": {
                        "description": "Registrations are disabled",
                        "schema": {
                            "$ref": "#/definitions/main.KError"
                        }
                    },
                    "409": {
                        "description": "Duplicated email or username",
                        "schema": {
                            "$ref": "#/definitions/main.KError"
                        }
                    },
                    "422": {
                        "description": "Invalid register body",
                        "schema": {
                            "$ref": "#/definitions/main.KError"
                        }
                    }
                }
            }
        },
...
        "/users/{id}/logo": {
            "get": {
                "security": [
                    {
                        "Jwt": [
                            "users.read"
                        ]
                    }
                ],
                "description": "Get a user's logo (manual upload if available, gravatar otherwise)",
                "produces": [
                    "image/*"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user logo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The id or username of the user",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "No gravatar image found for this user",
                        "schema": {
                            "$ref": "#/definitions/main.KError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main.Jwt": {
            "type": "object",
            "properties": {
                "token": {
                    "description": "The jwt token you can use for all authorized call to either keibi or other services.",
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWUsImlhdCI6MTUxNjIzOTAyMn0.KMUFsIDTnFmyG3nMiGM6H9FNFUROf3wh7SmqJp-QV30"
                }
            }
        },
        "main.KError": {
            "type": "object",
            "properties": {
                "details": {},
                "message": {
                    "type": "string",
                    "example": "No user found with this id"
                },
                "status": {
                    "type": "integer",
                    "example": 404
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Jwt": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "Token": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}